Added back 1.0.4

+++ invai-back/src/test/java/es/caib/invai/back/service/mapper/maintenance/development/role/RoleMapperTest.java
@@ -107,6 +107,6 @@ class RoleMapperTest {
 
     @Test
     void map_nullId_returnsNull() {
-        assertNull(mapper.map((Long) null));
+        assertNull(mapper.map(null));
     }
 }
