+++ invai-back/src/test/java/es/caib/invai/back/service/mapper/maintenance/general/commission/CommissionMapperTest.java
@@ -148,6 +148,6 @@ class CommissionMapperTest {
 
     @Test
     void map_nullId_returnsNull() {
-        assertNull(mapper.map((Long) null));
+        assertNull(mapper.map(null));
     }
 }
